Death Menu working version
When I packaged it, the locker wouldn't package because I think unity is smart and doesnt let you package purchased assets.
The locker I used is located in the asset we bought for the school. 

here is the path to the prefab
TirgamesAssets > SchoolScene > Models > Interior > Furniture > Prefabs > SchoolLocker01_2Old

Once you put that locker prefab in the scene, put these numbers in for position, rotation, and scale
Position
X:271  Y:97   Z:140
Rotation
X:0  Y:140  Z:0
Scale
X:120  Y:120  Z:120
This should perfectly position the locker in the menu/spotlight to work great.
If not then I can tweak it on your computer for you. Sorry I couldnt find a better way